adding process helper for xstate
